+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:57805/","mongoDBName":"jest"}
+{"mongoUri":"mongodb://127.0.0.1:58119/","mongoDBName":"jest"}
