+++ src/main/factories/controllers/survey/add-survey/add-survey-controller-factory.ts
@@ -1,5 +1,5 @@
 import { makeLogControllerDecorator } from '@main/factories/decorators/log-controller-decorator-factory'
-import { makeDbAddSurvey } from '@main/factories/usecases/add-survey/db-add-account-factory'
+import { makeDbAddSurvey } from '@main/factories/usecases/survey/add-survey/db-add-account-factory'
 import { AddSurveyController } from '@presentation/controller/survey/add-survey/add-survey-controller'
 import { Controller } from '@presentation/protocols'
 import { makeAddSurveyValidation } from './add-survey-validation-factory'
