+++ src/main/routes/login-routes.ts
@@ -1,6 +1,6 @@
-import { adaptRoute } from '@main/adapters/express/express-route-adapter'
-import { makeLoginController } from '@main/factories/controllers/login/login-controller-factory'
-import { makeSignUpController } from '@main/factories/controllers/signup/signup-controller-factory'
+import { adaptRoute } from '@main/adapters/express-route-adapter'
+import { makeLoginController } from '@main/factories/controllers/auth/login/login-controller-factory'
+import { makeSignUpController } from '@main/factories/controllers/auth/signup/signup-controller-factory'
 import { Router } from 'express'
 
 export default (router: Router): void => {
