+++ src/main/routes/survey-routes.ts
@@ -1,5 +1,5 @@
-import { adaptRoute } from '@main/adapters/express/express-route-adapter'
-import { makeAddSurveyController } from '@main/factories/controllers/add-survey/add-survey-controller-factory'
+import { adaptRoute } from '@main/adapters/express-route-adapter'
+import { makeAddSurveyController } from '@main/factories/controllers/survey/add-survey/add-survey-controller-factory'
 import { Router } from 'express'
 
 export default (router: Router): void => {
